// kaart/src/app/model/parameter-zoekobject.ts
export type ParameterZoekobject =
  | 'activiteit.naam'
  | 'activiteit.groep'
  | 'beperkingengebied.naam'
  | 'beperkingengebied.groep'
  | 'functie.naam'
  | 'functie.groep'
  | 'omgevingsnorm.naam'
  | 'omgevingsnorm.groep'
  | 'omgevingswaarde.naam'
  | 'omgevingswaarde.groep'
  | 'type'
  | 'thema'
  | 'documentIdentificatie';